Remove dead code from test.
